# Java/MoppyAPI/openapi.yaml
openapi: 3.0.0
info:
  title: Moppy API
  description: |
    REST API for controlling Moppy musical hardware devices.

    Moppy is a musical floppy drive controller that can play MIDI files through floppy drives, hard drives, and other computer hardware. This API provides programmatic control over MIDI playback, allowing you to load songs, control playback, and adjust parameters like tempo, volume, and position.

    ## Features
    - Load and play MIDI files (.mid, .midi)
    - Control playback (play, pause, stop)
    - Adjust tempo and volume in real-time
    - Seek to specific positions in songs
    - Loop playback
    - Get real-time playback status and progress
    - Monitor network status and connected devices
    - Discover Moppy devices on the network

    ## Getting Started
    1. Start the Moppy API server: `java -jar MoppyAPI-2.2.0.jar [--port 8080]`
    2. Check network status using `/api/network/status`
    3. View discovered devices using `/api/network/devices`
    4. Load a MIDI file using the `/api/load` endpoint
    5. Control playback using the play/pause/stop endpoints
    6. Monitor status using the `/api/state` or `/api/status` endpoints

  version: 2.2.0
  contact:
    name: Moppy Project
    url: https://github.com/SammyIAm/Moppy2
  license:
    name: MIT License
    url: https://github.com/SammyIAm/Moppy2/blob/main/LICENSE
servers:
  - url: http://localhost:8080
    description: Default local server
  - url: http://localhost:{port}
...
    variables:
...
paths:
...
  /api/load:
    post:
      tags:
        - playback
      summary: Load a MIDI file
      description: Load a MIDI file (.mid or .midi) for playback
      operationId: loadSong
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - filePath
              properties:
                filePath:
                  type: string
                  description: Absolute or relative path to the MIDI file
                  example: "/path/to/song.mid"
            examples:
              load_song:
                summary: Load a MIDI file
                value:
                  filePath: "/home/user/music/tetris.mid"
      responses:
        "200":
          description: Song loaded successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  value:
                    success: true
                    message: "Song loaded successfully"
                    fileName: "tetris.mid"
        "400":
          description: Bad request (invalid file path, unsupported format, or file not found)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                file_not_found:
                  value:
                    error: "MIDI file not found: /path/to/nonexistent.mid"
                    status: 400
                invalid_format:
                  value:
                    error: "File must be a MIDI file (.mid or .midi)"
                    status: 400
  /api/play:
    post:
      tags:
        - playback
      summary: Start playback
      description: Start or resume playback of the loaded MIDI file
      operationId: play
      responses:
        "200":
          description: Playback started successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  value:
                    success: true
                    message: "Playback started"
        "400":
          description: Bad request (no song loaded)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              examples:
                no_song:
                  value:
                    error: "No song loaded"
                    status: 400
        "405":
          description: Method not allowed (only POST is supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/pause:
    post:
      tags:
        - playback
      summary: Pause playback
      description: Pause the current playback (can be resumed with /api/play)
      operationId: pause
      responses:
        "200":
          description: Playback paused successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  value:
                    success: true
                    message: "Playback paused"
        "405":
          description: Method not allowed (only POST is supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/stop:
    post:
      tags:
        - playback
      summary: Stop playback
      description: Stop playback and reset position to beginning
      operationId: stop
      responses:
        "200":
          description: Playback stopped successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
              examples:
                success:
                  value:
                    success: true
                    message: "Playback stopped"
        "405":
          description: Method not allowed (only POST is supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/state:
    get:
      tags:
        - state
      summary: Get full playback state
      description: Get complete playback state information including all parameters and metadata
      operationId: getState
      responses:
        "200":
          description: Current playback state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PlaybackState"
        "405":
          description: Method not allowed (only GET is supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/status:
    get:
      tags:
        - state
      summary: Get playback status
      description: Get simplified playback status with formatted time information
      operationId: getStatus
      responses:
        "200":
          description: Current playback status
          content:
            application/json:
              schema:
                type: object
                properties:
                  state:
                    $ref: "#/components/schemas/PlaybackStateEnum"
                  isPlaying:
                    type: boolean
                    description: Whether playback is currently active
                    example: true
                  fileName:
                    type: string
                    description: Name of the loaded file
                    example: "tetris.mid"
                  progress:
                    type: number
                    format: double
                    minimum: 0
                    maximum: 1
                    description: Playback progress as a decimal (0.0 to 1.0)
                    example: 0.25
                  position:
                    type: string
                    description: Current position in MM:SS format
                    example: "1:30"
                  duration:
                    type: string
                    description: Total duration in MM:SS format
                    example: "6:00"
        "405":
          description: Method not allowed (only GET is supported)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/position:
    get:
      tags:
        - parameters
      summary: Get current position
      description: Get the current playback position in seconds
      operationId: getPosition
      responses:
        "200":
          description: Current position
          content:
            application/json:
              schema:
                type: object
                properties:
                  position:
                    type: integer
                    description: Current position in seconds
                    example: 90
                  formattedPosition:
                    type: string
                    description: Current position in MM:SS format
                    example: "1:30"
    put:
      tags:
        - parameters
      summary: Set playback position
      description: Seek to a specific position in the song
      operationId: setPosition
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - position
              properties:
                position:
                  type: integer
                  minimum: 0
                  description: Position in seconds to seek to
                  example: 120
            examples:
              seek_to_2min:
                summary: Seek to 2 minutes
                value:
                  position: 120
      responses:
        "200":
          description: Position set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  position:
                    type: integer
                    example: 120
        "400":
          description: Bad request (invalid position or missing parameter)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/tempo:
    get:
      tags:
        - parameters
      summary: Get current tempo
      description: Get the current playback tempo in BPM
      operationId: getTempo
      responses:
        "200":
          description: Current tempo
          content:
            application/json:
              schema:
                type: object
                properties:
                  tempo:
                    type: number
                    format: float
                    description: Current tempo in beats per minute
                    example: 120.0
    put:
      tags:
        - parameters
      summary: Set playback tempo
      description: Adjust the playback tempo (speed) of the song
      operationId: setTempo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - tempo
              properties:
                tempo:
                  type: number
                  format: float
                  minimum: 0.1
                  maximum: 1000.0
                  description: Tempo in beats per minute
                  example: 140.0
            examples:
              faster_tempo:
                summary: Set faster tempo
                value:
                  tempo: 140.0
              slower_tempo:
                summary: Set slower tempo
                value:
                  tempo: 80.0
      responses:
        "200":
          description: Tempo set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  tempo:
                    type: number
                    format: float
                    example: 140.0
        "400":
          description: Bad request (invalid tempo or missing parameter)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/loop:
    get:
      tags:
        - parameters
      summary: Get loop mode status
      description: Check if loop mode is enabled
      operationId: getLoop
      responses:
        "200":
          description: Current loop mode status
          content:
            application/json:
              schema:
                type: object
                properties:
                  loop:
                    type: boolean
                    description: Whether loop mode is enabled
                    example: false
    put:
      tags:
        - parameters
      summary: Set loop mode
      description: Enable or disable loop mode for continuous playback
      operationId: setLoop
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - loop
              properties:
                loop:
                  type: boolean
                  description: Whether to enable loop mode
                  example: true
            examples:
              enable_loop:
                summary: Enable loop mode
                value:
                  loop: true
              disable_loop:
                summary: Disable loop mode
                value:
                  loop: false
      responses:
        "200":
          description: Loop mode set successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  loop:
                    type: boolean
                    example: true
        "400":
          description: Bad request (missing parameter)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      description: Standard success response
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: "Operation completed successfully"
      additionalProperties: true
    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - error
        - status
      properties:
        error:
          type: string
          description: Error message
          example: "Invalid request"
        status:
          type: integer
          description: HTTP status code
          example: 400
...